chore: Add lazy loading for images and update global styles
This commit adds lazy loading for images by importing the necessary dependencies and updating the global styles. It also includes the addition of the i18n configuration file, updates to the tailwind.config.ts file, and updates to npm dependencies for i18next and lucide-react. Additionally, the ESLint configuration is updated to disable the react/prop-types rule. Finally, unused index.tsx file is removed and index.jsx is updated.

Note: This commit message follows the established convention of using a prefix (chore) to indicate a chore or maintenance task in the codebase.